move user model to userlogic
